Extract throttler config
