+++ src/app/config/app.config.ts
@@ -1,7 +1,4 @@
-import { ClassProvider } from '@nestjs/common';
 import { ConfigModuleOptions } from '@nestjs/config';
-import { APP_GUARD } from '@nestjs/core';
-import { minutes, seconds, ThrottlerGuard, ThrottlerModuleOptions } from '@nestjs/throttler';
 import * as Joi from 'joi';
 import { parseNumber } from '../utils';
 
@@ -50,22 +47,4 @@ export const configOptions: ConfigModuleOptions = {
         abortEarly: true,
         allowUnknowns: false,
     },
-};
-
-export const provideThrottlerGuard: () => ClassProvider = () => ({
-    provide: APP_GUARD,
-    useClass: ThrottlerGuard,
-});
-
-export const throttlerOptions: ThrottlerModuleOptions = {
-    throttlers: [
-        {
-            ttl: minutes(1),
-            limit: 100,
-        },
-        {
-            ttl: seconds(1),
-            limit: 10,
-        },
-    ],
 };
