+++ src/app/config/index.ts
@@ -1,1 +1,2 @@
 export * from './app.config';
+export * from './throttler.config';
